refactor: merge auth context and provider and add user webinarEnabled data
