+++ app/_shared/components/containers/app-container.tsx
@@ -1,19 +1,34 @@
 import NextUIProvider from '@/_shared/providers/nextui';
-import AuthProvider from '@/_shared/providers/auth';
-import type { UserType } from '@/_shared/types/user';
+import { AuthProvider } from '@/_shared/contexts/auth';
+import type { UserType } from '@/_shared/types/auth';
 import SignInModal from '@/_shared/components/auth/sign-in-modal';
+import { featureFlag } from '@/_shared/utils/feature-flag';
 
 export default function AppContainer({
   children,
   user,
 }: {
   children: React.ReactNode;
-  user: UserType.AuthUserData | null;
+  user: UserType.AuthUserResponse | null;
 }) {
   return (
     <>
       <NextUIProvider>
-        <AuthProvider value={{ user: user }}>
+        <AuthProvider
+          user={
+            user
+              ? {
+                  userID: user.id,
+                  email: user.email,
+                  name: user.name,
+                  pictureUrl: user.picture_url,
+                  // eslint-disable-next-line @typescript-eslint/ban-ts-comment
+                  // @ts-ignore
+                  webinarEnabled: user?.webinar_enabled === true,
+                }
+              : user
+          }
+        >
           {children}
           <SignInModal />
         </AuthProvider>
